# Improve the range evaluation for raw values without range index