@Jackie-Jiang 2022-05-18
Seems it can cause wrong result for floating point values because `1.0 + Double.MIN_VALUE == 1.0` (lose precision). For the inclusive range we might need to allow some delta for floating point values similar to equals