@Jackie-Jiang 2022-05-18
Solved with `Math.nextUp()` and `Math.nextDown()`